Faz 2:İlan güncelleme eklendi

+++ src/main/java/com/microcommerce/mini_coderspace/repository/JobPostingRepository.java
@@ -3,5 +3,5 @@ package com.microcommerce.mini_coderspace.repository;
 import com.microcommerce.mini_coderspace.entity.JobPosting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JobPostingRepository extends JpaRepository<JobPosting, Integer> {
+public interface JobPostingRepository extends JpaRepository<JobPosting, Long> {
 }
